+++ crates/hyperdrive-math/src/lib.rs
@@ -1,3 +1,4 @@
+mod base;
 mod long;
 mod lp;
 mod short;
