+++ crates/hyperdrive-math/src/lp/remove.rs
@@ -1,0 +1,6 @@
+use crate::State;
+
+impl State {}
+
+#[cfg(test)]
+mod tests {}
